Add files via upload
This project contains an implementation of RESTful architecture. It uses python's Django web framework in the back and also its (Django's) rest library. It uses ajax request to consume rest service to update a model called Person. It is a simple demonstration of what a rest architecture could be used for.